Refactor back to cosmos_db.py
