+++ tools/c7n_azure/tests/cassettes/CosmosDBTest.test_offer_collection.json
